Split body IR tests into separate files
